+++ lib/molecules/app_bar/app_bar_module.dart
@@ -1,1 +1,4 @@
-export 'app_bar.dart';
+export 'legacy/app_bar.dart';
+export 'legacy/profile_submenu_button.dart';
+export 'logo_app_bar.dart';
+export 'selection_button.dart';