+++ lib/molecules/molecules_module.dart
@@ -1,6 +1,7 @@
 export './table/listo_data_table.dart';
 export './tabs/tabs.dart';
-export 'app_bar/app_bar.dart';
+export 'app_bar/app_bar_module.dart';
+export 'app_bar/legacy/app_bar.dart';
 export 'card_list/card_list.dart';
 export 'navigation_menu/menu_icon.dart';
 export 'navigation_menu/navigation_menu.dart';