+++ widgets_demo/lib/components/layout/layout.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:widgetbook/widgetbook.dart';
+import 'package:widgets_demo/components/DEPRECATED/deprecated.dart';
 
-import 'cards/cards.dart';
 import 'scaffold/app_bar.dart';
 import 'tabs/tabs.dart';
 import 'theme/main_theme.dart';
@@ -13,6 +13,6 @@ WidgetbookCategory layoutCategory(BuildContext context) {
     scaffoldMolecule(context),
     titleLine(context),
     tabsMolecule(context),
-    cardsMolecule(context),
+    deprecatedCategory(context),
   ]);
 }